+++ RELEASENOTES.md
@@ -1,5 +1,12 @@
 # Release notes
 
+## 1.0.3
+- cmdline argument to scale B scan width added
+- app.py renamed to octlasagne.py
+
+## 1.0.2
+- annotation timer added
+
 ## 1.0.1
 - fix: wrong package name in setup.py
 - add: example data added